# book-service/src/main/java/library/dto/auth/LoginResponseDto.java
package library.dto.auth;

import library.entities.UserProfile;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.EqualsAndHashCode;
import lombok.NoArgsConstructor;

@Data
@EqualsAndHashCode(callSuper = false)
@NoArgsConstructor
@AllArgsConstructor
public class LoginResponseDto {
    private String token;
    private UserProfile profile;
}